requirements.txt
asgiref==3.11.0
boto3==1.42.16
botocore==1.42.16
brevo-python==1.2.0
certifi==2025.11.12
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==6.0
django-allauth==65.13.1
django-cleanup==9.0.0
django-crispy-forms==2.5
django-debug-toolbar==6.1.0
django-js-asset==3.1.2
django-storages==1.14.6
djangorestframework==3.16.1
jmespath==1.0.1
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
redis==7.1.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
urllib3==2.6.2
whitenoise==6.11.0
